fix: fix userModel
