+++ apps/web/src/schema/inventory/job-task/model.ts
@@ -1,27 +1,4 @@
-import type { Tags, TimeStamp } from '@/schema/_common/model';
-import type { USER_TYPE } from '@/schema/identity/user/constant';
-import type { RoleBindingModel } from '@/schema/inventory/cloud-service-query-set/model';
-
 import type { JobTaskError } from '@/services/asset-inventory/types/collector-history-job-type';
-
-export interface UserModel {
-    created_at: TimeStamp;
-    domain_id?: string;
-    email?: string;
-    email_verified?: boolean;
-    language: string;
-    last_accessed_at: TimeStamp | number;
-    name: string;
-    state: string;
-    tags?: Tags;
-    timezone: string;
-    user_id: string;
-    backend: string;
-    user_type: UserType;
-    role_bindings?: RoleBindingModel[];
-}
-
-export type UserType = typeof USER_TYPE[keyof typeof USER_TYPE];
 
 export interface JobTaskModel {
     job_task_id: string;
