+++ apps/web/src/services/administration/store/user-page-store.ts
@@ -4,7 +4,7 @@ import { SpaceConnector } from '@cloudforet/core-lib/space-connector';
 import type { Query } from '@cloudforet/core-lib/space-connector/type';
 
 import { USER_TYPE } from '@/schema/identity/user/constant';
-import type { UserType } from '@/schema/inventory/job-task/model';
+import type { UserType } from '@/schema/identity/user/type';
 import { store } from '@/store';
 
 import ErrorHandler from '@/common/composables/error/errorHandler';
